machines/biologic/edge/pyproject.toml
[project]
name = "biologic-edge"
version = "0.1.0"
description = "Edge service for the 'biologic' machine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easy-biologic==0.5.0rc1",
    "puda-comms",
    "python-dotenv>=1.2.1",
    "setuptools<81",
]

[tool.uv.sources]
puda-comms = {workspace=true}
